Pull out logic
